Abstract duplicate data store unit tests
