+++ tests/phpunit/jobstore/ActionScheduler_DBStore_Test.php
@@ -1,10 +1,30 @@
 <?php
 
+use Action_Scheduler\Tests\DataStores\AbstractStoreTest;
+
 /**
  * Class ActionScheduler_DBStore_Test
  * @group tables
  */
-class ActionScheduler_DBStore_Test extends ActionScheduler_UnitTestCase {
+class ActionScheduler_DBStore_Test extends AbstractStoreTest {
+
+	public function setUp() {
+		global $wpdb;
+
+		// Delete all actions before each test.
+		$wpdb->query( "DELETE FROM {$wpdb->actionscheduler_actions}" );
+
+		parent::setUp();
+	}
+
+	/**
+	 * Get data store for tests.
+	 *
+	 * @return ActionScheduler_DBStore
+	 */
+	protected function get_store() {
+		return new ActionScheduler_DBStore();
+	}
 
 	public function test_create_action() {
 		$time      = as_get_datetime_object();
@@ -442,94 +462,4 @@ class ActionScheduler_DBStore_Test extends ActionScheduler_UnitTestCase {
 		$this->assertEquals( ActionScheduler_Store::STATUS_FAILED, $store->get_status( $action_id ) );
 	}
 
-	/* Start tests for \ActionScheduler_DBStore::query_actions */
-
-	public function test_query_actions_query_type_arg_invalid_option() {
-		$this->expectException( InvalidArgumentException::class );
-
-		$store = new ActionScheduler_DBStore();
-		$store->query_actions( array( 'hook' => 'my_hook' ), 'invalid' );
-	}
-
-	public function test_query_actions_query_type_arg_valid_options() {
-		$this->delete_all_actions();
-
-		$store    = new ActionScheduler_DBStore();
-		$schedule = new ActionScheduler_SimpleSchedule( as_get_datetime_object( 'tomorrow' ) );
-
-		$action_id_1 = $store->save_action( new ActionScheduler_Action( 'my_hook', array( 1 ), $schedule ) );
-		$action_id_2 = $store->save_action( new ActionScheduler_Action( 'my_hook', array( 1 ), $schedule ) );
-
-		$this->assertEquals( array( $action_id_1, $action_id_2 ), $store->query_actions( array( 'hook' => 'my_hook' ) ) );
-		$this->assertEquals( 2, $store->query_actions( array( 'hook' => 'my_hook' ), 'count' ) );
-	}
-
-	public function test_query_actions_by_single_status() {
-		$this->delete_all_actions();
-
-		$store    = new ActionScheduler_DBStore();
-		$schedule = new ActionScheduler_SimpleSchedule( as_get_datetime_object( 'tomorrow' ) );
-
-		$this->assertEquals( 0, $store->query_actions( array( 'status' => ActionScheduler_Store::STATUS_PENDING ), 'count' ) );
-
-		$action_id_1 = $store->save_action( new ActionScheduler_Action( 'my_hook_1', array( 1 ), $schedule ) );
-		$action_id_2 = $store->save_action( new ActionScheduler_Action( 'my_hook_2', array( 1 ), $schedule ) );
-		$action_id_3 = $store->save_action( new ActionScheduler_Action( 'my_hook_3', array( 1 ), $schedule ) );
-		$store->mark_complete( $action_id_3 );
-
-		$this->assertEquals( 2, $store->query_actions( array( 'status' => ActionScheduler_Store::STATUS_PENDING ), 'count' ) );
-		$this->assertEquals( 1, $store->query_actions( array( 'status' => ActionScheduler_Store::STATUS_COMPLETE ), 'count' ) );
-	}
-
-	public function test_query_actions_by_array_status() {
-		$this->delete_all_actions();
-
-		$store    = new ActionScheduler_DBStore();
-		$schedule = new ActionScheduler_SimpleSchedule( as_get_datetime_object( 'tomorrow' ) );
-
-		$this->assertEquals(
-			0,
-			$store->query_actions(
-				array(
-					'status' => array( ActionScheduler_Store::STATUS_PENDING, ActionScheduler_Store::STATUS_RUNNING ),
-				),
-				'count'
-			)
-		);
-
-		$action_id_1 = $store->save_action( new ActionScheduler_Action( 'my_hook_1', array( 1 ), $schedule ) );
-		$action_id_2 = $store->save_action( new ActionScheduler_Action( 'my_hook_2', array( 1 ), $schedule ) );
-		$action_id_3 = $store->save_action( new ActionScheduler_Action( 'my_hook_3', array( 1 ), $schedule ) );
-		$store->mark_failure( $action_id_3 );
-
-		$this->assertEquals(
-			3,
-			$store->query_actions(
-				array(
-					'status' => array( ActionScheduler_Store::STATUS_PENDING, ActionScheduler_Store::STATUS_FAILED ),
-				),
-				'count'
-			)
-		);
-		$this->assertEquals(
-			2,
-			$store->query_actions(
-				array(
-					'status' => array( ActionScheduler_Store::STATUS_PENDING, ActionScheduler_Store::STATUS_COMPLETE ),
-				),
-				'count'
-			)
-		);
-	}
-
-	/* End tests for \ActionScheduler_DBStore::query_actions */
-
-
-	/**
-	 * Delete all actions from the DB store.
-	 */
-	protected function delete_all_actions() {
-		global $wpdb;
-		$wpdb->query( "DELETE FROM {$wpdb->actionscheduler_actions}" );
-	}
 }
